requirements.txt
psutil
anthropic
python-dotenv
flask
requests
watchdog